requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-bootstrap-form==3.4
django-heroku==0.3.1
django-stdimage==5.3.0
djangorestframework==3.10.3
mysqlclient==2.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==6.0.0
gunicorn==20.0.4
